requirements.txt
altair==5.5.0
appium-python-client>=2.11,<3
attrs==23.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==42.0.5
gitdb==4.0.12
GitPython==3.1.45
h11==0.14.0
idna==3.6
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.10.0
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyarrow==22.0.0
pycparser==2.22
pydeck==0.9.1
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
selenium==4.21.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2025.2
urllib3==1.26.18
urllib3-secure-extra==0.1.0
watchdog==6.0.0
webdriver-manager==4.0.2
wsproto==1.2.0
pytest==9.0.2
pytest-xdist==3.6.1
